requirements.txt
Pillow>=10.0
anthropic>=0.39.0
